ephemeral-multi-db-playground/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pymysql==1.1.0
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0
